+++ beep/structure/arbin.py
@@ -1,6 +1,8 @@
+import os
 from datetime import datetime
 
 import pytz
+import pandas as pd
 
 from beep.conversion_schemas import ARBIN_CONFIG
 from beep import logger
